+++ Sources/QRCodeReaderView.swift
@@ -126,7 +126,7 @@ final public class QRCodeReaderView: UIView, QRCodeReaderDisplayable {
   // MARK: - Scan Result Indication
 
   func startTimerForBorderReset() {
-    DispatchQueue.main.asyncAfter(deadline: DispatchTime.now() + .seconds(1)) {
+    DispatchQueue.main.asyncAfter(deadline: .now() + .seconds(1)) {
       if let ovl = self.overlayView as? ReaderOverlayView {
         ovl.overlayColor = .white
       }